+++ .pre-commit-config.yaml
@@ -5,6 +5,7 @@ repos:
       - id: flake8
         types:
           - python
+        args: ["--max-line-length=88"]
       - id: trailing-whitespace
       - id: end-of-file-fixer
       - id: check-yaml
