+++ calitp_feed_checker/__init__.py
@@ -1,15 +1,15 @@
 from collections import OrderedDict
 import json
-import typer
 import urllib.error
 import urllib.parse
 import urllib.request
 import yaml
 
-from transitland import get_transitland_urls
-from transitfeeds import get_transitfeeds_urls
+from .transitland import get_transitland_urls
+from .transitfeeds import get_transitfeeds_urls
 
 
+__version__ = "1.0.0"
 SECRET_PARAMS = ["api_key", "token", "apiKey", "key"]
 
 
@@ -26,12 +26,8 @@ def clean_url(url):
     return urllib.parse.urlunparse(url)
 
 
-def main(
-    yml_file=typer.Argument("agencies.yml", help="A yml file containing urls"),
-    csv_file=typer.Option(None, help="A csv file (one url per line)"),
-    url=typer.Option(None, help="URL to check instead of a file",),
-    output=typer.Option(None, help="Path to a file to save output to."),
-    verbose: bool = typer.Option(False, help="Print a result table to stdout"),
+def check_feeds(
+    yml_file=None, csv_file=None, url=None, output=None, verbose=False,
 ):
     results = {}
 
@@ -108,7 +104,3 @@ def main(
         with open(output, "w") as f:
             f.write(json.dumps(results, indent=4))
             print(f"Results saved to {output}")
-
-
-if __name__ == "__main__":
-    typer.run(main)
